app/pyproject.toml
[project]
name = "photo-attendance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.121.2",
    "insightface>=0.7.3",
    "opencv-python>=4.12.0.88",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.0",
    "uvicorn>=0.38.0",
]
